Move function and adjust name / psuedocode
Move function getComputerChoice to a new step 2 in the pseudocode so
that this function is just about getting a random number between 0 and
100. Change function name to getRandomNumber.

This is done so that the functions are clean and serve a single purpose.

In a later edit it may make more sense to condense functions, but as
per The Odin Project teachings its good to have functions serve one
purpose (as you can always callback functions within one connecting
function).