Add computerChoice variable to keep random number
The variable computerChoice will keep the random result from the
getRandomNumber function once, unless if the random number is 100 in
which case the first if conditional will rerun the getRandomNumber
function.

Before when the conditionals had getRandomNumber() it would
have kept creating new random numbers for each check (so it wouldn't
have worked as intended).