Adjust game variables and add console.log
Change the variables rock, paper, and scissors to equal themselves as
strings (first letter capitalized) as the check for values is done
within the conditional checks in getComputerChoice.

Fix the check for Rock by making computerChoice >= 0, not
computerChoice > 0 which results in the else being activated when the
number 0 is randomly generated by getRandomNumber.