Change conditional order and toLowerCase for fix
I will test to make sure this works, but this might fix the error.

In Chrome Dev Tools debugger the value for humanChoice did not
successfully get converted to all lowercase as per toLowerCase method,
so by integrating the toLowerCase method into the humanChoice declaration
hopefully this will fix it.